back-end/requirements.txt
Flask
Flask-Cors
torch
Pillow
gunicorn